api/services/auth/requirements.txt
uvicorn
fastapi
pydantic
psycopg2-binary
pyjwt
python-multipart